Split IconGenerator into focused classes
Extract ReadmeUpdater, RegistryGenerator, and NamingHelper from the
monolithic IconGenerator class. IconGenerator now only handles Blazor
component generation, improving single-responsibility and testability.
